refactor: 💡 refactor Keyboard related code to keyboard folder
